% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bvar.R
\name{bernoullirnd}
\alias{bernoullirnd}
\title{Generate sample from Bernoulli random variables}
\usage{
bernoullirnd(p)
}
\arguments{
\item{p}{vector of probabilities for zero}
}
\value{
the vector of 0/1 Bernoulli random variables
}
\description{
Generate sample from independent Bernoulli random variables with different probabilities
}
\details{
Generate sample from independent Bernoulli random variables with different probabilities
}
\examples{
probs <- c(0.5,0.8,0.1)
bvarr:::bernoullirnd(probs)
}

