% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bvar.imp.plot}
\alias{bvar.imp.plot}
\title{Draw irfs of bayesian VAR model}
\usage{
bvar.imp.plot(bvar.model, qus = c(0.1, 0.5, 0.9))
}
\arguments{
\item{bvar.model}{the list containing all results of bayesian VAR estimation}

\item{qus}{the vector of quantiles for irfs}
}
\value{
the ggplot2 graph of IRFs
}
\description{
Draw irfs of bayesian VAR model
}
\examples{
bvar(Yraw)
}

