% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bvar.pred.plot}
\alias{bvar.pred.plot}
\title{Estimate six types of bayesian VAR models}
\usage{
bvar.pred.plot(bvar.model)
}
\arguments{
\item{bvar.model}{the list containing all results of bayesian VAR estimation}
}
\value{
the list containing all results of bayesian VAR estimation
}
\description{
This function completes the subsetting, transforming and ordering triad
with a function that works in a similar way to \code{\link{subset}} and
\code{\link{transform}} but for reordering a data frame by its columns.
This saves a lot of typing!
}
\examples{
bvar(Yraw)
}

