% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvar_conjugate.R
\name{bvar_build_X}
\alias{bvar_build_X}
\title{Build X matrix from supplied data}
\usage{
bvar_build_X(Y_in, Z_in = NULL, constant = TRUE, p = 1)
}
\arguments{
\item{Y_in}{endogeneous variables [T_in x m]}

\item{Z_in}{optional exogeneous variables [T_in x ...]}

\item{constant}{whether the constant should be included, default is TRUE}

\item{p}{number of lags}
}
\value{
X [T x k] matrix of right hand side regressors: endogeneous and exogeneous variables,
T = T_in - p, k = m * p + d, d is the number of exogeneous variables including constant if present
X = [endo | const | other exo]
}
\description{
Build X matrix from supplied data
}
\details{
Build X matrix from supplied data
}
\examples{
data(Yraw)
X <- bvar_build_X(Yraw, constant = TRUE, p = 4)
}

