% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvar_conjugate.R
\name{bvar_build_Y}
\alias{bvar_build_Y}
\title{Build Y matrix from supplied data}
\usage{
bvar_build_Y(Y_in, p = 1)
}
\arguments{
\item{Y_in}{multivariate time series [T_in x m]}

\item{p}{number of lags}
}
\value{
Y [T x m] matrix of left hand side endogeneous variables, T = T_in - p
}
\description{
Build Y matrix from supplied data
}
\details{
Build Y matrix from supplied data
}
\examples{
data(Yraw)
Y <- bvar_build_Y(Yraw, p = 4)
}
