% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvar_conjugate.R
\name{bvar_conj_delta}
\alias{bvar_conj_delta}
\title{Create [m x 1] vector of deltas from delta description}
\usage{
bvar_conj_delta(Y_in, delta)
}
\arguments{
\item{Y_in}{multivariate time series}

\item{delta}{description of delta. May be: one number, "AR1", or ready to go [m x 1] vector}
}
\value{
[m x 1] vector of deltas
}
\description{
Create [m x 1] vector of deltas from delta description
}
\details{
Create [m x 1] vector of deltas from delta description
}
\examples{
data(Yraw)
dummy <- bvar_conj_delta(Yraw, delta = "AR1")
}
