% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvar_conjugate.R
\name{bvar_conj_estimate}
\alias{bvar_conj_estimate}
\title{Estimate bvar conjugate model from setup}
\usage{
bvar_conj_estimate(setup, keep = 2000, verbose = FALSE)
}
\arguments{
\item{setup}{list containing:
X [T x k]
Y [T x m]
X_plus [T_plus x k]
Y_plus [T_plus x m]
v_prior
p number of lags}

\item{keep}{the number of simulations.
If keep is zero only posterior hyperparameters are calculated}

\item{verbose}{TRUE will show some debugging messages, FALSE by default}
}
\value{
model list containing
X, Y, X_plus, Y_plus,
p, v_prior,
sample (if keep>0)
Omega_post, v_post, S_post, Phi_post
}
\description{
Estimate bvar conjugate model from setup
}
\details{
Estimate bvar conjugate model from setup
}
\examples{
data(Yraw)
setup <- bvar_conj_setup(Yraw, p = 4, lambda = c(0.2, 1, 1, 1, 100, 100))
model <- bvar_conj_estimate(setup, keep = 10)
}

