% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bvar_conjugate.R
\name{bvar_conj_lambda2dummy}
\alias{bvar_conj_lambda2dummy}
\title{Create dummy observations from lambdas}
\usage{
bvar_conj_lambda2dummy(Y_in, Z_in = NULL, constant = TRUE, p = 4,
  lambda = c(0.2, 1, 1, 1, 100, 100), delta = 1, s2_lag = NULL,
  y_bar_type = c("initial", "all"))
}
\arguments{
\item{Y_in}{multivariate time series}

\item{Z_in}{exogeneous variables}

\item{constant}{logical, default is TRUE, whether the constant should be included}

\item{p}{number of lags}

\item{lambda}{vector = (l_1, l_lag, l_sc, l_io, l_const, l_exo), the l_kron is set to 1 automatically for
conjugate N-IW prior. Short summary valid for NO sc/io case:
sd(const in eq i) = l_const * sigma_i
sd(exo in eq i)= l_exo * sigma_i
sd(coef for var j lag l in eq i) = l_1*sigma_i/sigma_j/l^l_lag
lambdas may be Inf
l_io or l_sc equal to NA means no corresponding dummy observations}

\item{delta}{vector [m x 1] or scalar or "AR1". Are used for prior Phi_1 and in sc/io dummy observations
Scalar value is replicated m times. If set to "AR1" then deltas will be estimated as AR(1) coefficients (but not greater than one).
Diagonal of Phi_1 is equal to delta. y_bar is multiplied by delta componentwise.}

\item{s2_lag}{number of lags in AR() model used to estimate s2 (equal to p by default)
Carriero uses 1 in his matlab code}

\item{y_bar_type}{(either "all" or "initial"). Determines how y_bar for sc and io dummy is calculated.
"all": y_bar is mean of y for all observations, "initial": p initial observations
Carriero: all, Sim-Zha: initial}
}
\value{
dummy list containing:
X_cniw,  Y_cniw
X_sc, Y_sc
X_io, Y_io
X_plus, Y_plus binding all corresponging Xs and Ys
}
\description{
Create dummy observations from lambdas
}
\details{
Create dummy observations from lambdas.
Lambdas specification is based on Carriero p. 52-53
}
\examples{
data(Yraw)
dummy <- bvar_conj_lambda2dummy(Yraw, p = 4, lambda = c(0.2,1,1,1,100,100))
}

