% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvar_conjugate.R
\name{bvar_conj_mdd}
\alias{bvar_conj_mdd}
\title{Calculate log marginal data density}
\usage{
bvar_conj_mdd(model)
}
\arguments{
\item{model}{estimated conjugate N-IW model}
}
\value{
log of marginal data density
}
\description{
Calculate log marginal data density
}
\details{
Calculate log marginal data density. Formula (22) from Carriero
"Bayesian VARs: Specification Choices and Forecast Accuracy" section 4.1.
}
\examples{
data(Yraw)
setup <- bvar_conj_setup(Yraw, p = 4)
model <- bvar_conj_estimate(setup = setup, keep = 100)
bvar_conj_mdd(model)
}

