% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bvar_conjugate.R
\name{bvar_conj_sigma2}
\alias{bvar_conj_sigma2}
\title{Create [m x 1] vector of sigma^2 from supplied time series}
\usage{
bvar_conj_sigma2(Y_in, s2_lag = 1)
}
\arguments{
\item{Y_in}{multivariate time series}

\item{s2_lag}{number of lags to estimate sigma^2}
}
\value{
[m x 1] vector of sigmas^2
}
\description{
Create [m x 1] vector of sigma^2 from supplied time series
}
\details{
Create [m x 1] vector of sigma^2 from supplied time series
}
\examples{
data(Yraw)
sigma2 <- bvar_conj_sigma2(Yraw)
}

