% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bvar_conjugate.R
\name{bvar_conj_simulate}
\alias{bvar_conj_simulate}
\title{Simulate from conjugate N-IW posterior distribution}
\usage{
bvar_conj_simulate(v_post, Omega_post_root, S_post, Phi_post, verbose = FALSE,
  keep = 10)
}
\arguments{
\item{v_post}{posterior nu, scalar}

\item{Omega_post_root}{[k x k] "root" of Omega posterior matrix}

\item{S_post}{[m x m] posterior scale matrix for noise covariance}

\item{verbose}{TRUE will show some debugging messages, FALSE by default}

\item{keep}{number of simulations}

\item{Phi_pose}{[k x m] posterior expected values for coefs}
}
\value{
coda object with simulations
}
\description{
Simulate from conjugate N-IW posterior distribution
}
\details{
Simulate from conjugate N-IW posterior distribution
}
\examples{
data(Yraw)
dummy <- bvar_conj_lambda2dummy(Yraw,p=2)
hyper <- bvar_dummy2hyper(dummy$Y_cniw, dummy$X_cniw)
# these are priors but just for testing let's pretend that they are posteriors
post_sample <- bvar_conj_simulate(v_post=10,
hyper$Omega_root, hyper$S, hyper$Phi)
}

