% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bvar_conjugate.R
\name{bvar_conj_summary}
\alias{bvar_conj_summary}
\title{summary of a conjugate Normal-Inverse-Wishart bayesian VAR model}
\usage{
bvar_conj_summary(model)
}
\arguments{
\item{model}{estimated conjugate N-IW model}
}
\value{
nothing
}
\description{
summary of a conjugate Normal-Inverse-Wishart bayesian VAR model
}
\details{
summary of a conjugate Normal-Inverse-Wishart bayesian VAR model
}
\examples{
data(Yraw)
setup <- bvar_conj_setup(Yraw, p = 4)
model <- bvar_conj_estimate(setup = setup, keep=100)
bvar_conj_summary(model)
}

