% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bvar_conjugate.R
\name{bvar_conjugate_posterior}
\alias{bvar_conjugate_posterior}
\title{Calculate posterior hyperparameters from X_star and Y_star matrix}
\usage{
bvar_conjugate_posterior(Y_star, X_star)
}
\arguments{
\item{Y_star}{[T_star x m] matrix of left hand side endogeneous variables}

\item{X_star}{[T_star x k] matrix of right hand side regressors: endogeneous and exogeneous variables}
}
\description{
Calculate posterior hyperparameters from X_star and Y_star matrix
}
\details{
Calculate posterior hyperparameters from X_star and Y_star matrix.
Usually X_star and Y_star are equal to X and Y augmented with dummy observations.
}
\examples{
data(Yraw)
X <- bvar_build_X(Yraw, constant=TRUE, p=4)
Y <- bvar_build_Y(Yraw, p=4)
post <- bvar_conjugate_posterior(Y,X) # no dummy observation, just for demo
}

