% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bvar_conjugate.R
\name{bvar_dummy2hyper}
\alias{bvar_dummy2hyper}
\title{Calculate hyperparameters from artificial observations}
\usage{
bvar_dummy2hyper(Y_star, X_star)
}
\arguments{
\item{Y_star}{[T_star x m] matrix of left hand side endogeneous variables}

\item{X_star}{[T_star x k] matrix of right hand side regressors: endogeneous and exogeneous variables}
}
\value{
list of hyperparameters: Omega, Omega_root, Phi, S
}
\description{
Calculate hyperparameters from artificial observations
}
\details{
Calculate hyperparameters from artificial observations. Can be used both for prior and posterior
hyperparameters. In the case of prior hyperparameters X_star usually consists of X_cniw, X_io and X_sc.
In the case of posterior hyperparameters X_star usually consists of X, X_cniw, X_io and X_sc.
}
\examples{
data(Yraw)
X <- bvar_build_X(Yraw, constant=TRUE, p=4)
Y <- bvar_build_Y(Yraw, p=4)
post <- bvar_dummy2hyper(Y,X) # no dummy observation, just for demo
}

