% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bvar_conjugate.R
\name{bvar_get_endo_varnames}
\alias{bvar_get_endo_varnames}
\title{Get endogeneous variable names from supplied data}
\usage{
bvar_get_endo_varnames(Y_in)
}
\arguments{
\item{Y_in}{[T_in x m] multivariate time series}
}
\value{
vector of variable names
}
\description{
Get endogeneous variable names from supplied data
}
\details{
Get endogeneous variable names from supplied data
}
\examples{
data(Yraw)
bvar_get_endo_varnames(Yraw)
}

