% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvar_conjugate.R
\name{bvar_get_exo_varnames}
\alias{bvar_get_exo_varnames}
\title{Get exogeneous variable names from supplied data}
\usage{
bvar_get_exo_varnames(Z_in, constant = TRUE)
}
\arguments{
\item{Z_in}{[T_in x m] multivariate time series}

\item{constant}{logical, default is TRUE, whether the constant should be included}
}
\value{
vector of variable names
}
\description{
Get exogeneous variable names from supplied data
}
\details{
Get exogeneous variable names from supplied data
}
\examples{
data(Yraw)
bvar_get_exo_varnames(Yraw)
}

