% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bvar_conjugate.R
\name{bvar_get_Y_in}
\alias{bvar_get_Y_in}
\title{Recover original Y_in from Y, X and number of lags p}
\usage{
bvar_get_Y_in(Y, X, p)
}
\arguments{
\item{Y}{[T x m] matrix of left hand side endogeneous variables}

\item{X}{[T x k] matrix of right hand side regressors: endogeneous and exogeneous variables}

\item{p}{number of lags}
}
\value{
original supplied Y_in
}
\description{
Recover original Y_in from Y, X and number of lags p
}
\details{
Recover original Y_in from Y, X and number of lags p
}
\examples{
data(Yraw)
Y <- bvar_build_Y(Yraw, p=4)
X <- bvar_build_X(Yraw, constant=TRUE, p=4)
Y_in <- bvar_get_Y_in(Y,X,p=4)
}

