% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bvar_conjugate.R
\name{bvar_hyper2dummy}
\alias{bvar_hyper2dummy}
\title{Calculate artificial observations from hyperparameters}
\usage{
bvar_hyper2dummy(Omega, S, Phi)
}
\arguments{
\item{Omega}{[k x k] hyperparameter, scale matrix for covariance of coefficients}

\item{S}{[m x m] hyperparameter, scale matrix for IW-distribution}

\item{Phi}{[k x m] hyperparameter, expected values of coefficients}
}
\value{
list of artificial observations: X_plus, Y_plus
}
\description{
Calculate artificial observations from hyperparameters
}
\details{
Calculate artificial observations from hyperparameters.
Usually used to calculate artificial observations
from prior hyperparameters.
}
\examples{
data(Yraw)
X <- bvar_build_X(Yraw, constant=TRUE, p=4)
Y <- bvar_build_Y(Yraw, p=4)
hyper <- bvar_dummy2hyper(Y,X) # no dummy observation, just for demo
dummy <- bvar_hyper2dummy(hyper$Omega, hyper$S, hyper$Phi)
}

