% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bvar_conjugate.R
\name{Carriero_priors}
\alias{Carriero_priors}
\title{Set conjugate N-IW priors from lambdas as in Carriero}
\usage{
Carriero_priors(Y_in, Z_in = NULL, constant = TRUE, p = 4,
  lambdas = c(1, 0.2, 1, 1), VAR_in = c("levels", "growth rates"))
}
\arguments{
\item{Y_in}{multivariate time series}

\item{Z_in}{exogeneous variables}

\item{p}{number of lags}

\item{lambdas}{vector = (l0, l1, l3, l4), the l2 is set to 1 automatically for
conjugate N-IW prior}

\item{VAR_in}{(either "levels" or "growth rates")}
}
\value{
priors list containing Phi_prior [k x m], Omega_prior [k x k], S_prior [m x m], v_prior [1x1],
where k = mp+d
}
\description{
Set conjugate N-IW priors from lambdas as in Carriero
}
\details{
Set conjugate N-IW priors from lambdas as in Carriero
Based on Carriero p. 52-53
}
\examples{
data(Yraw)
priors <- Carriero_priors(Yraw, p = 4, lambdas = c(1,0.2,1,1))
model <- bvar_conjugate0(priors = priors)
}

