% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_forecast.R
\name{evaluate_bvar}
\alias{evaluate_bvar}
\title{Evaluate density forecast with point realisations}
\usage{
evaluate_bvar(rolling_bvar, plot = T, include_point = T)
}
\arguments{
\item{rolling_bvar}{object of class `rolling_bvar`, see \link[rolling_BVAR]{bvarr}}

\item{plot}{produce forecast evaluation plots}

\item{include_point}{include point forecast accuracy (uses median of predictive density)}
}
\value{
Value of the score
}
\description{
Evaluate density forecast with point realisations
}
\details{
crps_sample employs an empirical version of the quantile
  decomposition of the CRPS (Laio and Tamea, 2007) when using method = "edf".
  For method = "kde", it uses kernel density estimation using a Gaussian
  kernel. The logarithmic score always uses kernel density estimation. The
  bandwidth (bw) for kernel density estimation can be specified manually, in
  which case it must be a positive number. If bw == NULL, the bandwidth is
  selected using the core function bw.nrd. Numerical integration may speed up
  computation for crps_sample in case of large samples dat.

  A lower score indicates a better forecast
}
\examples{
df <- ggplot2::economics[,c(1, 4:6)]
bvar_res <- rolling_BVAR(df, date, start = "2013-01-01", h = 1, fixedWindow = F,
                          n_draws = 2000,
                          lags = 4,
                          cores = 3)
evaluate_bvar(bvar_res)

}
\references{
Evaluating simulation based forecast distributions: Krueger, F., Lerch, S.,
Thorarinsdottir, T.L. and T. Gneiting (2016): 'Probabilistic forecasting and
comparative model assessment based on Markov Chain Monte Carlo output',
working paper, Heidelberg Institute for Theoretical Studies, available at
http://arxiv.org/abs/1608.06802.

Empirical quantile decomposition of the CRPS: Laio, F. and S. Tamea (2007):
'Verification tools for probabilistic forecasts of continuous hydrological
variables', Hydrology and Earth System Sciences, 11, 1267-1277.
}
