% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bvar_conjugate.R
\name{forecast_conjugate}
\alias{forecast_conjugate}
\title{predict with conjugate Normal-Inverse-Wishart bayesian VAR model}
\usage{
forecast_conjugate(model, Y_in = NULL, Z_f = NULL, useMean = FALSE,
  h = 1, level = c(80, 95), type = c("prediction", "credible"))
}
\arguments{
\item{model}{estimated conjugate N-IW model}

\item{Y_in}{(NULL by default) past values of endogeneous variables (shold have at least p observations).
If NULL, then Y_in supplied for estimation will be used. Only last p values of Y_in are used.}

\item{Z_f}{future values of exogeneous variables}

\item{useMean}{(FALSE by default) use mean values (TRUE) of posterior draws of y_t,
or median values (FALSE)}

\item{h}{number of periods for forecasting}

\item{level}{confidence levels for prediction intervals}

\item{type}{("prediction" by default) type of interval: "prediction" incorporates uncertainty about
future shocks; "credible" deals only with parameter uncertainty.}
}
\value{
forecast results
}
\description{
predict with conjugate Normal-Inverse-Wishart bayesian VAR model
}
\details{
predict with conjugate Normal-Inverse-Wishart bayesian VAR model
}
\examples{
data(Yraw)
priors <- lambda2priors(Yraw, p = 4, lambdas = c(1,0.2,1,1,1))
model <- bvar_conjugate0(Yraw, p = 4, priors = priors)
forecst_conjugate(model)
}

