% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bvar_conjugate.R
\name{KK_code_priors}
\alias{KK_code_priors}
\title{Set conjugate N-IW priors as in matlab code of Koops-Korobilis}
\usage{
KK_code_priors(Y, Z = NULL, constant = TRUE, p = 4)
}
\arguments{
\item{Y}{multivariate time series}

\item{Z}{exogeneous variables}

\item{p}{number of lags}
}
\value{
priors list containing Phi_prior [k x m], Omega_prior [k x k], S_prior [m x m], v_prior [1x1],
where k = mp+d
}
\description{
Set conjugate N-IW priors as in matlab code of Koops-Korobilis
}
\details{
Set conjugate N-IW priors as in matlab code of Koops-Korobilis
}
\examples{
data(Yraw)
priors <- KK_code_priors(Yraw, p = 4)
model <- bvar_conjugate0(Yraw, p = 4, priors = priors)
}

