% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bvar_conjugate.R
\name{lambda2priors}
\alias{lambda2priors}
\title{Set conjugate N-IW priors from lambdas}
\usage{
lambda2priors(Y, Z = NULL, constant = TRUE, p = 4, lambdas = c(1, 0.2,
  1, 1, 1), VAR_in = c("levels", "growth rates"))
}
\arguments{
\item{Y}{multivariate time series}

\item{Z}{exogeneous variables}

\item{p}{number of lags}

\item{lambdas}{vector = (l0, l1, l2, l3, l4)}

\item{VAR_in}{(either "levels" or "growth rates")}
}
\value{
priors list containing Phi_prior [k x m], Omega_prior [k x k], S_prior [m x m], v_prior [1x1],
where k = mp+d
}
\description{
Set conjugate N-IW priors from lambdas
}
\details{
Set conjugate N-IW priors from lambdas
Based on Carriero p. 53
}
\examples{
data(Yraw)
priors <- lambda2priors(Yraw, p = 4, lambdas = c(1,0.2,1,1,1))
model <- bvar_conjugate0(Yraw, p = 4, priors = priors)
}

