% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvar_conjugate.R
\name{lmvgamma}
\alias{lmvgamma}
\title{Multivariate log-gamma-function}
\usage{
lmvgamma(p, a)
}
\arguments{
\item{p}{dimension}

\item{a}{argument}
}
\value{
log G_p(a), log of multivariate gamma-function
}
\description{
Multivariate log-gamma-function
 
Multivariate log-gamma-function. Wikipedia-style parametrisation (p,a)
}
\examples{
lmvgamma(2,3)
}

