% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvar.R
\docType{data}
\name{macro_russia}
\alias{macro_russia}
\title{Russian macroeconomic indicators data.frame}
\format{A data.frame with 243 rows and 24 variables:
\describe{
  \item{construction}{construction}
  \item{cpi}{cpi}
  \item{employment}{employment}
  \item{export}{export}
  \item{gas_price}{gas_price}
  \item{gov_balance}{gov_balance}
  \item{ib_rate}{ib_rate}
  \item{import}{import}
  \item{ind_prod}{ind_prod}
  \item{labor_request}{labor_request}
  \item{lend_rate}{lend_rate}
  \item{m2}{m2}
  \item{ner}{ner}
  \item{nfa_cb}{nfa_cb}
  \item{oil_price}{oil_price}
  \item{ppi}{ppi}
  \item{real_income}{real_income}
  \item{real_investment}{real_investment}
  \item{reer}{reer}
  \item{retail}{retail}
  \item{time_y}{time_y}
  \item{unemp_rate}{unemp_rate}
  \item{wage}{wage}
  \item{agriculture}{agriculture}
}}
\source{
\url{http://www.cbr.ru/eng/}
}
\usage{
macro_russia
}
\description{
Monthly russian data on various macroeconomic indicators. 
1995:M1 - 2015:M3. The variables are as follows:
}
\keyword{datasets}

