% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bvar.R
\name{makeblock}
\alias{makeblock}
\title{makeblock}
\usage{
makeblock(X, i, p)
}
\arguments{
\item{X}{matrix}

\item{i}{the number of block to generate}

\item{p}{number of lags, is equal to total the number of blocks}
}
\value{
block
}
\description{
makeblock
}
\details{
used only inside mlag2 function
}
\examples{
X <- matrix(1:18, ncol=3)
makeblock(X,1,2)
}

