% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvar_conjugate.R
\name{marginal_data_density}
\alias{marginal_data_density}
\title{Calculate log marginal data density}
\usage{
marginal_data_density(model)
}
\arguments{
\item{model}{estimated conjugate N-IW model}
}
\value{
log of marginal data density
}
\description{
Calculate log marginal data density
}
\details{
Calculate log marginal data density. Formula (22) from Carriero
"Bayesian VARs: Specification Choices and Forecast Accuracy" section 4.1.
}
\examples{
data(Yraw)
priors <- Carriero_priors(Yraw, p = 4)
model <- bvar_conjugate0(priors = priors, fast_forecast = TRUE)
marginal_data_density(model)
}
