% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_BVAR.R
\name{plot.rolling_bvar}
\alias{plot.rolling_bvar}
\title{Plot object of class rolling_bvar}
\usage{
\method{plot}{rolling_bvar}(rolling_bvar, plot = T)
}
\description{
plots the distribution and line plots for the forecasts using ggplot
}
