% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_BVAR.R
\name{rolling_BVAR}
\alias{rolling_BVAR}
\title{Rolling/expanding Bayesian VAR estimate}
\usage{
rolling_BVAR(df, date_col, start, h = 1, fixedWindow = T, include = "raw",
  n_draws = 2000, lags = 4, cores = 2)
}
\arguments{
\item{df}{data.frame containg Y variables and a date column}

\item{date_col}{specify column name which contains t time elements}

\item{start}{start date from which the rolling estimation should begin ex. "2010-01-01"}

\item{h}{what should the forecast horison be (Now only h = 1, by = 1)}

\item{fixedWindow}{whether it should be expaning or rolling. Fixed = T is rolling}

\item{include}{keep raw for sample analysis}

\item{n_draws}{number of draws to keep}

\item{lags}{number of lags for VAR estimate}

\item{cores}{if cores are specified then forecasts are implemented in parallel}
}
\value{
data.frame
}
\description{
Performs a rolling/expanding estimation of BVAR forecast estimates with re-estimation at each point t. Later implementation will use a model opbject
}
\examples{

df <- ggplot2::economics[,c(1, 4:6)]


bvar_res <- rolling_BVAR(df, date, start = "2013-01-01", by = 1, fixedWindow = F,
                  n_draws = 2000,
                  lags = 4,
                  cores = 3)

bvar_res$forecast_density \%>\% filter(date == "2015-02-01") \%>\%
  tidyr::gather(., metric, value, -date) \%>\%
  ggplot(., aes(value)) +
  geom_density() +
  facet_wrap(~metric, scales = c("free"))

plot(bvar_res)

}
