% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvar_conjugate.R
\name{summary_conjugate}
\alias{summary_conjugate}
\title{summary of a conjugate Normal-Inverse-Wishart bayesian VAR model}
\usage{
summary_conjugate(model)
}
\arguments{
\item{model}{estimated conjugate N-IW model}
}
\value{
nothing
}
\description{
summary of a conjugate Normal-Inverse-Wishart bayesian VAR model
}
\details{
summary of a conjugate Normal-Inverse-Wishart bayesian VAR model
}
\examples{
data(Yraw)
priors <- Carriero_priors(Yraw, p = 4)
model <- bvar_conjugate0(priors = priors, keep = 100)
summary_conjugate(model)
}

