% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bvar_conjugate.R
\name{szbvar_priors}
\alias{szbvar_priors}
\title{Set conjugate N-IW priors from lambdas and mus as in Sim Zha}
\usage{
szbvar_priors(Y_in, Z_in = NULL, constant = TRUE, p = 4, lambdas = c(1,
  0.2, 1, 1, 1, 1), mu56 = c(1, 1), VAR_in = c("levels", "growth rates"))
}
\arguments{
\item{Y_in}{multivariate time series}

\item{Z_in}{exogeneous variables}

\item{constant}{logical, default is TRUE, whether the constant should be included}

\item{p}{number of lags}

\item{lambdas}{vector = (l0, l1, l3, l4, l5), l2 is set to 1 for conjugate N-IW}

\item{mu56}{vector = (mu5, mu6)}

\item{VAR_in}{(either "levels" or "growth rates")}
}
\value{
priors list containing Phi_prior [k x m], Omega_prior [k x k], S_prior [m x m], v_prior [1x1],
where k = mp+d
}
\description{
Set conjugate N-IW priors from lambdas and mus as in Sim Zha
}
\details{
Set conjugate N-IW priors from lambdas and mus as in Sim Zha
Should be compatible with szbvar function. Maybe error!!!!
MAYBE lambda should be in denominator!!!! Article vs MSBVAR???
}
\examples{
data(Yraw)
priors <- szbvar_priors(Yraw, p = 4, lambdas = c(1,0.2,1,1,1), mu56=c(1,1))
model <- bvar_conjugate0(priors = priors, keep=100)
}

