% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bvar.imp.plot}
\alias{bvar.imp.plot}
\title{Draw irfs of bayesian VAR model}
\usage{
bvar.imp.plot(bvar.model, qus = c(0.1, 0.5, 0.9))
}
\arguments{
\item{bvar.model}{the list containing all results of bayesian VAR estimation}

\item{qus}{the vector of quantiles for irfs}
}
\value{
the list containing all results of bayesian VAR estimation
}
\description{
This function completes the subsetting, transforming and ordering triad
with a function that works in a similar way to \code{\link{subset}} and
\code{\link{transform}} but for reordering a data frame by its columns.
This saves a lot of typing!
}
\examples{
bvar(Yraw)
}

