% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bvar.pred.plot}
\alias{bvar.pred.plot}
\title{Histograms of forecasts}
\usage{
bvar.pred.plot(bvar.model)
}
\arguments{
\item{bvar.model}{the list containing all results of bayesian VAR estimation}
}
\value{
the ggplot graph of forecast histograms
}
\description{
This function draws histograms of forecasts
}
\examples{
bvar(Yraw)
}

