% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{bvar}
\alias{bvar}
\alias{bvar-package}
\title{bvar}
\usage{
bvar(Yraw, prior = "SSVS-SSVS", W = NULL, p = 4, constant = TRUE,
  nsave = 10000, nburn = 2000, it_print = 2000, impulses = TRUE,
  ihor = 24, forecasting = TRUE, repfor = 50, h = 1)
}
\arguments{
\item{Yraw}{the matrix or data.frame with endogeneous VAR variables}

\item{prior}{the type of prior: "diffuse","Minnesota","conjugate", "independent","SSVS-Wishart","SSVS-SSVS"}

\item{W}{the matrix of exogenous variables}

\item{p}{the number of lags for endogeneous variables}

\item{constant}{(TRUE/FALSE) indicator, whether the constant is included}

\item{nsave}{the length of mcmc chain}

\item{nburn}{the length of burn-in of mcmc chain}

\item{it_print}{which iteration are printed}

\item{impulses}{whether to calculate irfs}

\item{ihor}{the number of lags for irfs}

\item{forecasting}{whether to calculate forecasts}

\item{repfor}{number of times to obtain a draw from the predictive}

\item{h}{number of forecast periods}
}
\value{
the list containing all results of bayesian VAR estimation
}
\description{
bvar
}
\examples{
bvar(Yraw)
}
\author{
Gary Koop, Korobilis, Boris Demeshev
Estimate six types of bayesian VAR models

This function completes the subsetting, transforming and ordering triad
with a function that works in a similar way to \code{\link{subset}} and
\code{\link{transform}} but for reordering a data frame by its columns.
This saves a lot of typing!
}

