% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bvar.summary}
\alias{bvar.summary}
\title{Print summary for bayesian VAR model}
\usage{
bvar.summary(bvar.model)
}
\arguments{
\item{bvar.model}{the list containing all results of bayesian VAR estimation}
}
\description{
This function completes the subsetting, transforming and ordering triad
with a function that works in a similar way to \code{\link{subset}} and
\code{\link{transform}} but for reordering a data frame by its columns.
This saves a lot of typing!
}
\examples{
bvar(Yraw)
}

