% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bvar.summary}
\alias{bvar.summary}
\title{Print summary for bayesian VAR model}
\usage{
bvar.summary(bvar.model)
}
\arguments{
\item{bvar.model}{the list containing all results of bayesian VAR estimation}
}
\description{
Print summary for bayesian VAR model
}
\examples{
bvar(Yraw)
}

