% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bvar.R
\name{bvar.summary}
\alias{bvar.summary}
\title{Print summary for bayesian VAR model}
\usage{
bvar.summary(model)
}
\arguments{
\item{bvar.model}{the list containing all results of bayesian VAR estimation}
}
\description{
Print summary for bayesian VAR model
}
\examples{
data(Yraw)
model <- bvar(Yraw,prior = "independent")
bvar.summary(model)
}

