% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bvar_conjugate.R
\name{bvar_conjugate0}
\alias{bvar_conjugate0}
\title{Estimate conjugate Normal-Inverse-Wishart bayesian VAR model}
\usage{
bvar_conjugate0(Y_in = NULL, Z_in = NULL, constant = TRUE, p = NULL,
  keep = 10000, verbose = FALSE, priors = list(Phi_prior = NULL,
  Omega_prior = NULL, S_prior = NULL, v_prior = NULL, Y_dummy = NULL, X_dummy =
  NULL, Y_in = NULL, Z_in = NULL, p = NULL), fast_forecast = FALSE)
}
\arguments{
\item{Y_in}{the matrix or data.frame with endogeneous VAR variables}

\item{Z_in}{(NULL by default) the matrix or data.frame with exogeneous VAR variables}

\item{constant}{(TRUE by default) whether we should include constant}

\item{p}{(2 by default) the number of lags}

\item{keep}{(10000 by default) the number of Gibbs sampling replications to keep
Is ignored when the fast_forecast is TRUE.}

\item{verbose}{(FALSE by default)}

\item{priors}{the list containing Phi_prior [k x m], Omega_prior [k x k],
S_prior [m x m], v_prior [1x1],
Y_dummy [T_dummy x m], X_dummy [T_dummy x k]
where k = mp+d}

\item{fast_forecast}{logical, FALSE by default. If TRUE then no simulations are done,
only posterior hyperparameters are calculated.}
}
\value{
the list containing all results of bayesian VAR estimation
}
\description{
Estimate conjugate  Normal-Inverse-Wishart bayesian VAR model
}
\details{
Estimate conjugate  Normal-Inverse-Wishart bayesian VAR model
}
\examples{
data(Yraw)
priors <- Carriero_priors(Yraw, p = 4)
model <- bvar_conjugate0(priors = priors)
}

