% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bvar_conjugate.R
\name{forecast_conjugate}
\alias{forecast_conjugate}
\title{predict with conjugate Normal-Inverse-Wishart bayesian VAR model}
\usage{
forecast_conjugate(model, Y_in = NULL, Z_f = NULL, output = c("long",
  "wide"), h = 1, level = c(80, 95), type = c("prediction", "credible"),
  out_of_sample = TRUE, include = c("mean", "median", "sd"))
}
\arguments{
\item{model}{estimated conjugate N-IW model}

\item{Y_in}{(NULL by default) past values of endogeneous variables (shold have at least p observations).
If NULL, then Y_in supplied for estimation will be used. Only last p values of Y_in are used.}

\item{Z_f}{future values of exogeneous variables}

\item{output}{(default "long") --- long or wide table for mean/quantiles of forecasts}

\item{h}{number of periods for forecasting}

\item{level}{confidence levels for prediction intervals}

\item{type}{("prediction" by default) type of interval: "prediction" incorporates uncertainty about
future shocks; "credible" deals only with parameter uncertainty.}

\item{out_of_sample}{logical, default is TRUE, whether forecasts are out of sample or not.
If forecasts are not out of sample, then parameter h is ignored}

\item{include}{(default is c("mean", "median", "sd")) what type of summary to provide
If include is NULL and level is NULL then the function will return raw mcmc predictions}
}
\value{
forecast results
}
\description{
predict with conjugate Normal-Inverse-Wishart bayesian VAR model
}
\details{
predict with conjugate Normal-Inverse-Wishart bayesian VAR model
}
\examples{
data(Yraw)
priors <- Carriero_priors(Yraw, p = 4)
model <- bvar_conjugate0(priors = priors)
forecast_conjugate(model, h=2, output="wide")
forecast_conjugaet(model, out_of_sample = FALSE, include="mean", level=NULL, type = "credible")
}

