% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bvar.R
\name{impulse}
\alias{impulse}
\title{impulse}
\usage{
impulse(By, smat, nstep)
}
\arguments{
\item{By}{[neq x nvar x nlag] array. neq=nvar, of course, but the first index runs over
equations. In response, the first index runs over variables, the second over
shocks (in effect, equations).}

\item{smat}{square matrix of initial shock vectors. To produce "orthogonalized
impulse responses" it should have the property that smat'*smat=sigma, where sigma
is the Var(u(t)) matrix and u(t) is the residual vector.  One way to get such a smat
is to set smat=chol(sigma).  To get the smat corresponding to a different ordering,
use smat=chol(P*Sigma*P')*P, where P is a permutation matrix.}

\item{nstep}{number of steps}
}
\value{
nvar x neq x nstep array
}
\description{
impulse
}
\details{
function response=impulse(By,smat,nstep), C. Sims' code.
}
\examples{
bvarr:::impulse(array(1, dim=c(2,2,2)), diag(2), 5)
}

