% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bvar.R
\name{mlag2}
\alias{mlag2}
\title{mlag2}
\usage{
mlag2(X, p)
}
\arguments{
\item{X}{matrix}

\item{p}{number of lags, is equal to total the number of blocks}
}
\value{
...
}
\description{
mlag2
}
\details{
Binds blocks provided by makeblock function
}
\examples{
X <- matrix(1:18, ncol=3)
mlag2(X,2)
}

