% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvar_datasets.R
\docType{data}
\name{Yraw}
\alias{Yraw}
\title{US inflation, employement and interest rate data.frame}
\format{A data.frame with 215 rows and 3 variables:
\describe{
  \item{inflation}{inflation}
  \item{employment}{employment}
  \item{interest_rate}{interest_rate}
}}
\source{
\url{https://sites.google.com/site/dimitriskorobilis/matlab/code-for-vars}
}
\usage{
Yraw
}
\description{
Quarterly US data on inflation, unemployment and interest rate,
1953:Q1 - 2006:Q3 from KK-code. The variables are as follows:
}
\keyword{datasets}
