% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{Yraw}
\alias{Yraw}
\title{US inflation, employement data.fram}
\format{A data frame with ... rows and .. variables}
\usage{
data(Yraw)
}
\description{
A dataset containing . The variables are as follows:
}
\details{
\itemize{
\item inflation. basdf sdjk
\item employment. blablba
\item .... sdfsd fsdlk flksdj f
}
}
\keyword{datasets}

