% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{Yraw}
\alias{Yraw}
\title{US inflation, employement and interest rate data.frame}
\format{A data frame with ... rows and .. variables}
\usage{
data(Yraw)
}
\description{
Quarterly US data on inflation, unemployment and interest rate,
1953:Q1 - 2006:Q3 The variables are as follows:
}
\details{
\itemize{
\item inflation. basdf sdjk
\item employment. blablba
\item interest_rate sdfsd fsdlk flksdj f
}
}
\keyword{datasets}

